Moving wrapLoadDevice back to session
