+++ lib/session.js
@@ -8,7 +8,6 @@ import * as hdnodeUtils from './utils/hdnode';
 import * as signTxHelper from './utils/signtx';
 import * as signBjsTxHelper from './utils/signbjstx';
 import {CallHelper} from './utils/call';
-import wrapLoadDevice from './utils/wrapLoadDevice';
 
 import * as trezor from './trezortypes';
 import type {TxInfo} from './utils/signbjstx';
@@ -357,4 +356,23 @@ export function coinNetwork(
         throw new Error(`No network with the name ${name}.`);
     }
     return network;
+}
+
+function wrapLoadDevice(
+    settings: trezor.LoadDeviceSettings,
+    network?: bitcoin.Network = bitcoin.networks.bitcoin
+): trezor.LoadDeviceSettings {
+    if (settings.node == null && settings.mnemonic == null) {
+        if (settings.payload == null) {
+            throw new Error('Payload, mnemonic or node necessary.');
+        }
+        try { // try to decode as xprv
+            const bjsNode = bitcoin.HDNode.fromBase58(settings.payload, network);
+            settings = { ...settings, node: hdnodeUtils.bjsNode2privNode(bjsNode) };
+        } catch (e) { // use as mnemonic
+            settings = { ...settings, mnemonic: settings.payload };
+        }
+        delete settings.payload;
+    }
+    return settings;
 }
